+++ src/main/java/com/example/springbootboardjpa/controller/PostController.java
@@ -1,21 +1,14 @@
 package com.example.springbootboardjpa.controller;
 
-import com.example.springbootboardjpa.dto.request.PostCreateRequest;
-import com.example.springbootboardjpa.dto.request.PostUpdateRequest;
-import com.example.springbootboardjpa.dto.response.PostResponse;
+import com.example.springbootboardjpa.dto.post.request.PostCreateRequest;
+import com.example.springbootboardjpa.dto.post.request.PostUpdateRequest;
+import com.example.springbootboardjpa.dto.post.response.PostResponse;
 import com.example.springbootboardjpa.service.PostService;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/posts")
